+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added legal notice and privacy declaration [#768](https://github.com/utetrapp/geli/issues/768)
 - Add more tests notification ctrl [#772](https://github.com/utetrapp/geli/pull/772)
 - Added more EN/DE translation [#753](https://github.com/utetrapp/geli/issues/753)
+- Simple e2e test for login [#795](https://github.com/utetrapp/geli/pull/795)
 - Added checkboxes for accepting our terms of use and privacy declarations whil registering [#778](https://github.com/utetrapp/geli/issues/778)
 - Quickfixed the Import/Export [#42](https://github.com/utetrapp/geli/issues/42)
 - Added PDF download of Coursecontent [#720](https://github.com/utetrapp/geli/pull/720)